// moneylaundering-web/src/app/main/case-info-requirement/models/news.model.ts
export class News {
    id: number;
    businessUnitId: number;
    caseId: number;
    newsTypeId: number;
    statusId: number;
    createDate: Date;
    newsTypeDescription: string;
    statusDescription: string;
    endDate: Date;
    expirationDate: Date;
    comments: string;
    newsReasonTypeId: number;
}